Products Page Data Base
This java file reads this data:
http://data.gc.ca/data/en/dataset/7ed8d72e-71e9-478e-8d0b-173b67a0752a

parses the csv file and gets: the lower age bound, higher age bound,
name of the products bought by this age group, and values of the
products bought by this age group.

This parsed database is suposed to be in the sqlite db on heroku, where
the php script in the backend grabs it and puts it on our mobile web
app. But we might now have time to finish this.